requirements.txt
numpy==1.23.3
pandas==1.5.0
sqlalchemy==1.4.41
psycopg2==2.9.3
google-api-python-client==2.63.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
csvkit==1.0.7
